+++ Makefile
@@ -28,7 +28,7 @@ MAKEFLAGS += --no-print-directory
 
 INCLUDES := -L $(LIBFT_DIR) -lft -lreadline
 
-#SRC += main.c \
+SRC += main.c \
 
 # ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ PARSING ⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒⌒ #
 SRC += $(addprefix $(DIR_TOKE), $(SRC_TOKE))
@@ -45,7 +45,8 @@ SRC_EXPAND:=	expansion.c \
 				expansion_utils.c \
 				letter_count.c \
 				word_count.c \
-				test/main_expand.c \
+				replace_var_env.c
+#test/main_expand.c \
 
 SRC +=  $(addprefix $(DIR_PARSE), $(SRC_PARSE))
 DIR_PARSE:= parsing/
